Perform requirement only if the server is actually present.
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@42950 a97e2381-89e5-4abb-bab3-167db6db766c

+++ tests/plugins/org.jboss.tools.ui.bot.ext/src/org/jboss/tools/ui/bot/ext/config/requirement/RemoveServer.java
@@ -17,8 +17,10 @@ public class RemoveServer extends RequirementBase {
 	@Override
 	public void handle() {
 		if (!checkFulfilled()) {
-			SWTTestExt.servers.deleteServer(SWTTestExt.configuredState.getServer().name);
-			SWTTestExt.eclipse.removeServerRuntime(SWTTestExt.configuredState.getServer().name);
+			if (SWTTestExt.servers.serverExists(SWTTestExt.configuredState.getServer().name)){
+				SWTTestExt.servers.deleteServer(SWTTestExt.configuredState.getServer().name);
+				SWTTestExt.eclipse.removeServerRuntime(SWTTestExt.configuredState.getServer().name);
+			}
 			SWTTestExt.configuredState.getServer().isConfigured=false;
 			SWTTestExt.configuredState.getServer().name=null;
 			SWTTestExt.configuredState.getServer().version=null;
